Add option for overriding structured data attributes
In the current implementation, only certain attributes can be set for
the structured data. This makes it impossible to use certain types like
"SoftwareApplication", which need unsupported attributes for google to
accept them for rich results. 😖

This change allows the user to override attributes in the structured
data tree by recursively merging in attributes from the
custom_structured_data subtree.