Truncate the description value to 280 chars
Currently the value of the `description`, `og: description`
and `twitter: description` tags includes the whole text of the post.

Which means that, on post pages, the text of the post is included *four*
times in the HTML document – adding to the page weight needlessly.

This commit ensures the description string is truncated to a sane value.